+++ biz/webui/htdocs/src/js/composer-list.js
@@ -1,5 +1,4 @@
 var React = require('react');
-var $ = require('jquery');
 var Composer = require('./composer');
 var util = require('./util');
 var LazyInit = require('./lazy-init');
@@ -21,24 +20,12 @@ var ComposerList = React.createClass({
     events.on('comTabsChange', function () {
       self.setState({});
     });
-    events.on('showMaskIframe', function () {
-      self.showMask();
-    });
-    events.on('hideMaskIframe', function () {
-      self.hideMask();
-    });
   },
   showTab: function (name) {
     if (this.state.activeName !== name) {
       storage.set('activeComposerTab', name);
       this.setState({ activeName: name });
     }
-  },
-  hideMask: function () {
-    $('.w-mask-iframe').hide();
-  },
-  showMask: function () {
-    $('.w-mask-iframe').show();
   },
   render: function () {
     var self = this;
@@ -105,11 +92,6 @@ var ComposerList = React.createClass({
           hide={hide}
           tabs={tabs}
           className="w-custom-tab-panel"
-        />
-        <div
-          className="w-mask-iframe"
-          onClick={this.hideMask}
-          onMouseEnter={this.hideMask}
         />
       </div>
     );