+++ biz/webui/htdocs/src/js/data-center.js
@@ -31,9 +31,9 @@ var hashFilterObj;
 var clearNetwork;
 var inited;
 var logId;
-var uploadFiles;
 var port;
 var pageId;
+var account;
 var dumpCount = 0;
 var updateCount = 0;
 var MAX_UPDATE_COUNT = 4;
@@ -59,10 +59,6 @@ exports.MAX_INCLUDE_LEN = MAX_INCLUDE_LEN;
 exports.MAX_EXCLUDE_LEN = MAX_EXCLUDE_LEN;
 exports.changeLogId = function (id) {
   logId = id;
-};
-
-exports.getUploadFiles = function () {
-  return uploadFiles;
 };
 
 exports.getPort = function () {
@@ -430,10 +426,7 @@ exports.values = createCgiObj(
     },
     add: 'cgi-bin/values/add',
     remove: 'cgi-bin/values/remove',
-    rename: 'cgi-bin/values/rename',
-    upload: 'cgi-bin/values/upload',
-    checkFile: 'cgi-bin/values/check-file',
-    removeFile: 'cgi-bin/values/remove-file'
+    rename: 'cgi-bin/values/rename'
   },
   POST_CONF
 );
@@ -591,6 +584,10 @@ exports.getComTabs = function () {
   return comTabList;
 };
 
+exports.getAccount = function() {
+  return account;
+};
+
 exports.getInitialData = function (callback) {
   if (!initialDataPromise) {
     initialDataPromise = $.Deferred();
@@ -600,12 +597,13 @@ exports.getInitialData = function (callback) {
         if (!data) {
           return setTimeout(load, 1000);
         }
-        port = data.server && data.server.port;
+        var server = data.server;
+        port = server && server.port;
+        account = server && server.account;
         updateCertStatus(data);
         exports.supportH2 = data.supportH2;
         exports.custom1 = data.custom1;
         exports.custom2 = data.custom2;
-        uploadFiles = data.uploadFiles;
         initialData = data;
         pageId = data.clientId;
         DEFAULT_CONF.data.clientId = pageId;
@@ -797,7 +795,9 @@ function startLoadData() {
       if (!data || data.ec !== 0) {
         return;
       }
-      port = data.server && data.server.port;
+      var server = data.server;
+      port = server && server.port;
+      account = server && server.account;
       updateCertStatus(data);
       exports.supportH2 = data.supportH2;
       exports.custom1 = data.custom1;
@@ -1425,7 +1425,10 @@ exports.getPlugin = function (name) {
 };
 
 function getMenus(menuName) {
-  var list = [];
+  var list = account && account[menuName];
+  if (!Array.isArray(list)) {
+    list = [];
+  }
   if (disabledAllPlugins) {
     return list;
   }