+++ biz/webui/htdocs/src/js/index.js
@@ -22,7 +22,7 @@ var storage = require('./storage');
 var Dialog = require('./dialog');
 var ListDialog = require('./list-dialog');
 var FilterBtn = require('./filter-btn');
-var FilesDialog = require('./files-dialog');
+// var FilesDialog = require('./files-dialog');
 var message = require('./message');
 var UpdateAllBtn = require('./update-all-btn');
 var ContextMenu = require('./context-menu');
@@ -766,11 +766,6 @@ var Index = React.createClass({
     });
     events.on('disableAllPlugins', self.disableAllPlugins);
     events.on('disableAllRules', self.disableAllRules);
-    events.on('showFiles', function (_, data) {
-      self.files = self.files || data;
-      self.showFiles();
-    });
-
     events.on('activeRules', function () {
       var rulesModal = dataCenter.rulesModal;
       if (rulesModal.exists(dataCenter.activeRulesName)) {
@@ -2710,9 +2705,9 @@ var Index = React.createClass({
   composer: function () {
     events.trigger('composer');
   },
-  showFiles: function () {
-    this.refs.filesDialog.show(this.files);
-  },
+  // showFiles: function () {
+  //   this.refs.filesDialog.show(this.files);
+  // },
   clear: function () {
     var modal = this.state.network;
     this.setState({
@@ -2730,10 +2725,16 @@ var Index = React.createClass({
         if (!sure) {
           return;
         }
-        dataCenter.rules.remove({ name: name }, function (data, xhr) {
+        var wholeGroup = sure === 2;
+        dataCenter.rules.remove({ name: name, wholeGroup: wholeGroup ? 1 : undefined }, function (data, xhr) {
           if (data && data.ec === 0) {
-            var nextItem = item && !item.active ? null : modal.getSibling(name);
-            modal.remove(name);
+            var nextItem;
+            if (wholeGroup) {
+              nextItem = modal.removeGroup(name);
+            } else {
+              nextItem = item && !item.active ? null : modal.getSibling(name);
+              modal.remove(name);
+            }
             if (nextItem) {
               self.setRulesActive(nextItem.name);
               events.trigger('expandRulesGroup', nextItem.name);
@@ -2751,7 +2752,7 @@ var Index = React.createClass({
             util.showSystemError(xhr);
           }
         });
-      });
+      }, util.isGroup(name));
     }
   },
   removeValues: function (item) {
@@ -2764,10 +2765,16 @@ var Index = React.createClass({
         if (!sure) {
           return;
         }
-        dataCenter.values.remove({ name: name }, function (data, xhr) {
+        var wholeGroup = sure === 2;
+        dataCenter.values.remove({ name: name, wholeGroup: wholeGroup ? 1 : undefined }, function (data, xhr) {
           if (data && data.ec === 0) {
-            var nextItem = item && !item.active ? null : modal.getSibling(name);
-            modal.remove(name);
+            var nextItem;
+            if (wholeGroup) {
+              nextItem = modal.removeGroup(name);
+            } else {
+              nextItem = item && !item.active ? null : modal.getSibling(name);
+              modal.remove(name);
+            }
             if (nextItem) {
               self.setValuesActive(nextItem.name);
               events.trigger('expandValuesGroup', nextItem.name);
@@ -2779,7 +2786,7 @@ var Index = React.createClass({
             util.showSystemError(xhr);
           }
         });
-      });
+      }, util.isGroup(name));
     }
   },
   setRulesActive: function (name, modal) {
@@ -3902,7 +3909,7 @@ var Index = React.createClass({
             isNetwork={isNetwork}
             hide={isPlugins}
           />
-          <a
+          {/* <a
             onClick={this.showFiles}
             className="w-files-menu"
             draggable="false"
@@ -3988,6 +3995,14 @@ var Index = React.createClass({
               className="w-help-menu-item"
             />
           </div>
+          <a
+            onClick={this.showHttpsSettingsDialog}
+            className="w-account-menu"
+            draggable="false"
+          >
+            <span className="glyphicon glyphicon-user"></span>
+            Account
+          </a>
           <Online name={name} />
           <div
             onMouseDown={this.preventBlur}
@@ -4727,7 +4742,7 @@ var Index = React.createClass({
             </button>
           </div>
         </Dialog>
-        <FilesDialog ref="filesDialog" />
+        {/* <FilesDialog ref="filesDialog" /> */}
         <ListDialog
           ref="selectRulesDialog"
           name="rules"